Bug fix for edge case
In the edge case that the cache is 128 bytes and split, there is
in theory no need for an index bit. In practice this means that
tag_bits and block_offset_bits add up to 32. This sum is used
to right-shift the index as much as needed after first left-
shifting it by tag_bits. The problem with this is that index
is a 32-bit integer, and shifting it is only defined when the
shift count is 0-31. So shifting by 32 is undefined, and its
behaviour can't be predicted. Therefore an if statement was
added to deal with this case.